ISOIML-28 make it work using systemd daemons and log to journalctl
Thanks to using SystemD we have automatic restarting of the service.
Because the programs are using publisher/subscriber pattern, they do not
have to synchronize message sending. All the configuration regarding the
services will be stored in the
/etc/systemd/system/<service_name>.service file.

The crash analysis will be performed by analyzing the journalctl, which
is cool :D